requirements.txt
ollama>=0.2.0
chromadb>=0.5.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
streamlit>=1.35.0
pandas>=2.0.0
pytest>=8.0.0
